requirements.txt
requests
feedparser
python-dotenv
pytest
pyyaml
